+++ lib/Steps/Integrations/__tests__/ReactNative.ts
@@ -1,6 +1,7 @@
 jest.mock('../../../Helper/Logging.ts'); // We mock logging to not pollute the output
 import * as fs from 'fs';
 import { Answers } from 'inquirer';
+import * as path from 'path';
 import * as process from 'process';
 import * as rimraf from 'rimraf';
 
@@ -10,8 +11,10 @@ import { ReactNative } from '../ReactNative';
 const testDir = 'rn-test';
 const iosIndexJs = 'index.ios.js';
 const appTsx = 'src/App.tsx';
+const appBuildGradle = 'android/app/build.gradle';
 
 const dummyJsContent = 'import React from "react";\n';
+const dummyAppBuildGradleContent = 'apply plugin: "com.facebook.react"\n\nandroid {\n}\n';
 
 const testArgs = {
   debug: false,
@@ -23,8 +26,17 @@ const testArgs = {
   url: 'https://not.used',
 };
 
-const testAnswers: Answers = {
+const mockIosAnswers: Answers = {
   shouldConfigurePlatforms: { 'ios': true },
+  config: {
+    dsn: {
+      public: 'dns.public.com',
+    },
+  },
+};
+
+const mockAndroidAnswers: Answers = {
+  shouldConfigurePlatforms: { 'android': true },
   config: {
     dsn: {
       public: 'dns.public.com',
@@ -41,8 +53,10 @@ describe('ReactNative', () => {
     fs.mkdirSync(testDir);
     process.chdir(testDir);
     fs.writeFileSync(iosIndexJs, dummyJsContent);
-    fs.mkdirSync('src');
+    fs.mkdirSync(path.dirname(appTsx), { recursive: true });
     fs.writeFileSync(appTsx, dummyJsContent);
+    fs.mkdirSync(path.dirname(appBuildGradle), { recursive: true });
+    fs.writeFileSync(appBuildGradle, dummyAppBuildGradleContent);
   });
 
   afterEach(() => {
@@ -52,7 +66,7 @@ describe('ReactNative', () => {
 
   test('patches js files', async () => {
     const project = new ReactNative(testArgs as Args);
-    await project.emit(testAnswers);
+    await project.emit(mockIosAnswers);
 
     const patchedIosIndexJs = fs.readFileSync(iosIndexJs, 'utf8');
     const patchedAppTsx = fs.readFileSync(appTsx, 'utf8');
@@ -63,5 +77,16 @@ describe('ReactNative', () => {
       '});\n\n';
     expect(patchedIosIndexJs).toEqual(expectedPatch);
     expect(patchedAppTsx).toEqual(expectedPatch);
+  });
+
+  test('patches android app build gradle file', async () => {
+    const project = new ReactNative(testArgs as Args);
+    await project.emit(mockAndroidAnswers);
+
+    const patchedAppBuildGradle = fs.readFileSync(appBuildGradle, 'utf8');
+    const expectedPatch = 'apply plugin: "com.facebook.react"\n\n' +
+      'apply from: "../../node_modules/@sentry/react-native/sentry.gradle"\n' +
+      'android {\n}\n';
+    expect(patchedAppBuildGradle).toEqual(expectedPatch);
   });
 });